# housingserver-specification/src/main/resources/openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080/api/v2/housing
    description: Local development server
info:
  description: Housing benefit application handling demo API
  version: "1.0.0"
  title: Housing benefit api API
  contact:
    email: riku.sarlin@gmail.com
  license:
    name: MIT
    url: 'https://mit-license.org/'
...
paths:
...
  /person/{id}:
    get:
      tags:
        - developers
      summary: fetch person
      operationId: fetchPersonById
      description: |
        Fetch person by person id
      parameters:
        - in: path
          name: id
          description: id of person to return
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: person found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: bad input parameter
        '404':
          description: person not found
    put:
      tags:
        - developers
      summary: update person
      operationId: updatePerson
      description: Updates person info
      parameters:
        - in: path
          name: id
          description: id of person to update
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: person updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'

        '400':
          description: update not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
        description: Person to update
    delete:
      tags:
        - admins
      summary: delete person
      operationId: deletePerson
      description: Deletes person
      parameters:
        - in: path
          name: id
          description: id of person to delete
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '204':
          description: person deleted successfully
        '400':
          description: delete not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: person not found
  /person/{id}/check:
    get:
      tags:
        - developers
      summary: cross-check person info
      operationId: checkPersonById
      description: |
        Cross validate by person id
      parameters:
        - in: path
          name: id
          description: id of person to check
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: person validation ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: person validation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: person not found
  /person:
    post:
      tags:
        - admins
      summary: add person
      operationId: addPerson
      description: Adds a person to the system
      responses:
        '201':
          description: person created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: person creation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '409':
          description: an existing person already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
        description: Person to add
...
components:
  schemas:
    ErrorMessage:
      type: object
      description: Info about error(s) in backend processing
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: number
          format: integer
          description: HTTP status code
        error:
          type: string
          description: description of error
        path:
          type: string
          description: URL path of error
        message:
          type: array
          items:
            type: object
            properties:
              fieldName:
                type: string
                description: name of field, if applicable, where error occured
              errorMessage:
                type: string
                description: a more detailed description of the error
              errorCode:
                type: string
                description: error code
    Id:
      type: integer
      description: Unique identifier of a resource
...
    Person:
      type: object
      description: A natural person residing in Finland
      properties: 
        id:
          $ref: '#/components/schemas/Id'
        personNumber:
          $ref: '#/components/schemas/PersonNumber'
        firstName:
          type: string
          description: First name of a person
          maxLength: 80
        lastName:
          type: string
          description: Last name of a person
          maxLength: 80
        birthDate:
          type: string
          format: date
          example: '1971-08-71'
...
    PersonNumber:
      type: string
      description: Finnish person number
      minLength: 11
      maxLength: 11
      pattern: '^\d{6}[A+-]\d{3}[0-9A-FHJ-NPR-Y]$'
...